LayoutTests/imported/w3c/web-platform-tests/webcodecs/full-cycle-test.https.any_h264_avc-expected.txt

CONSOLE MESSAGE: Error: assert_unreached: H264 AVC format is not yet supported Reached unreachable code

FAIL Encoding and decoding cycle promise_test: Unhandled rejection with value: object "ReferenceError: Can't find variable: OffscreenCanvas"
